deleteuser

+++ UI/Program.cs
@@ -13,18 +13,18 @@ internal class Program
         UserService userservise = new(identifier, userdb);
         MessageDB messageDB = new();
         MessageService messageService = new(messageDB);
-        
+
         //1. SKAPAKONTO
 
-          //user = CreateUser(user, logInService, userdb, identifier);
-          //userservise.MakeUser(userdb, user);
-          
-         
-               // DELETE FUNKAR EJ, VAAAAD ÄR KNAAAAAS??????
+        //user = CreateUser(user, logInService, userdb, identifier);
+        //userservise.MakeUser(userdb, user);
+
+
+        // DELETE FUNKAR EJ, VAAAAD ÄR KNAAAAAS??????
 
         //2. LOGGA IN PÅ BEFINTLIGT KONTO
         user = new();
-        user.Email = "angelinaholmqvist@live.se";//ConsoleInput.GetString("Enter your Email");
+        user.Email = "";//ConsoleInput.GetString("Enter your Email");
         user.Password = 1010;//ConsoleInput.GetInt("Enter your Password");
         user = logInService.UserLogIn(user); //user skriver bara i sin mail och kod
         user.Id = logInService.UserLogInIsValid(user); //andvänder userhandler och ser om user finns
@@ -33,7 +33,7 @@ internal class Program
             Console.WriteLine("Fel lösen eller mail");
             Environment.Exit(0);
         }
-    
+
         //1. TESTA GÖRA ANNONS
 
         //2. TESTA SÖKA ANNONS
@@ -44,9 +44,9 @@ internal class Program
         //500 kr
         // - visningsnamn
         // annonsid 
-        
-         //3. SKRIV MEDDELANDE TILL ANNONSEN
-       // int advertiseId = ConsoleInput.GetInt("Enter advertise ID to write message: ");
+
+        //3. SKRIV MEDDELANDE TILL ANNONSEN
+        // int advertiseId = ConsoleInput.GetInt("Enter advertise ID to write message: ");
         // int advertiseUserId = 11 ;//= userdb.getuserid(advertiseId);
         // Message message = new("KATTEN", "Jag vill gärna köpa din katt!", user.Id, advertiseUserId);
         // messageService.MakeMessage(message);
@@ -54,36 +54,36 @@ internal class Program
 
 
         // VISA ALLA MEDDELANDEN 
-        Message message = new();
-        user.messages = messageService.ShowAllMessages(user);
-        if (user.messages.Count() == 0)
-        {
-            Console.WriteLine("No Messages");
-        }
-        foreach (Message item in user.messages)
-        {
-            Console.WriteLine(item.MessagesToString());
-        }
-        // VÄLJ MEDDELANDE ATT LÄSA
-        int messageId = ConsoleInput.GetInt("Enter message to read: ");
-        Message readMessage = messageService.ShowOneMessage(messageId);
-        Console.WriteLine(readMessage.WholeMessageToString());
+        // Message message = new();
+        // user.messages = messageService.ShowAllMessages(user);
+        // if (user.messages.Count() == 0)
+        // {
+        //     Console.WriteLine("No Messages");
+        // }
+        // foreach (Message item in user.messages)
+        // {
+        //     Console.WriteLine(item.MessagesToString());
+        // }
+        // // VÄLJ MEDDELANDE ATT LÄSA
+        // int messageId = ConsoleInput.GetInt("Enter message to read: ");
+        // Message readMessage = messageService.ShowOneMessage(messageId);
+        // Console.WriteLine(readMessage.WholeMessageToString());
 
-        //4. SVARA PÅ MEDDELANDE
-        int chocie = ConsoleInput.GetInt("1 för att svara, 2 för att tillbaka");
-        if(chocie == 1)
-        {
-            string rubric = ConsoleInput.GetString("Rubric: ");
-            string content = ConsoleInput.GetString("Content: ");
-            int idToUser = readMessage.IDFromUser;
-            Message answerMessage = new(rubric, content, user.Id, idToUser);
-            messageService.MakeMessage(answerMessage);
-            Console.WriteLine("Skickat");
-        }
+        // //4. SVARA PÅ MEDDELANDE
+        // int chocie = ConsoleInput.GetInt("1 för att svara, 2 för att tillbaka");
+        // if(chocie == 1)
+        // {
+        //     string rubric = ConsoleInput.GetString("Rubric: ");
+        //     string content = ConsoleInput.GetString("Content: ");
+        //     int idToUser = readMessage.IDFromUser;
+        //     Message answerMessage = new(rubric, content, user.Id, idToUser);
+        //     messageService.MakeMessage(answerMessage);
+        //     Console.WriteLine("Skickat");
+        // }
         //5. REDIGERA PROFIL
         // DELETE USER
-           //  DeleteAUser(user,userdb);
-        //  userservise.DeleteTheUser(userdb, user); 
+        DeleteAUser(user, userdb);
+        userservise.DeleteTheUser(userdb, user);
 
         //6. VISA MINA ANNONSER
 
@@ -189,7 +189,7 @@ internal class Program
                 {
                     case "1":
                         //Raderar användare om användare finns
-                        
+
                         string delete = ConsoleInput.GetString(" ");
                         if (userHandeler.DeleteUser(user) == true)
                         {
@@ -240,7 +240,7 @@ internal class Program
             Environment.Exit(0);
         }
         user.SocialSecurityNumber = ConsoleInput.GetString("social security number: ");
-        if(identifier.ValidateSocialSecurityNumber(user.SocialSecurityNumber) == false)
+        if (identifier.ValidateSocialSecurityNumber(user.SocialSecurityNumber) == false)
         {
             Console.WriteLine("Social security number incorrect");
             Environment.Exit(0);
@@ -251,8 +251,16 @@ internal class Program
     }
     public static void DeleteAUser(User user, IUserHandeler userHandeler)
     {
-        user.Email = ConsoleInput.GetString("User mail: ");
+        string answer = ConsoleInput.GetString("Are you sure you want to delete your account? [yes] [no]");
+        if( answer == "yes")
+        {
         userHandeler.DeleteUser(user);
+        Console.WriteLine("Account deleted!");
+        }
+        else
+        {
+           Environment.Exit(0);
+        }
     }
 
     public static void ShowOneMessage(int messageId, IMessageHandeler messageHandeler) //A
@@ -292,7 +300,7 @@ internal class Program
         string municipality = string.Empty;
         int postalNumber = 0;
         User user = new();
-    
+
         advertise nyannons = new advertise(rubric, description, price, location, municipality, postalNumber, user.Id);
         return nyannons;
     }
